Add visual leader maps for diffget and diffput
Since d is the delete prefix, you can't use it for visual selections.
Create leader maps that do the same thing to make it easier to move
regions between diff files.

Unique so users will be notified of conflicts. Not silent since there's
nothing to squelch (want to see command and error).